Allow unpadded secrets
As specified
[here](https://github.com/google/google-authenticator/wiki/Key-Uri-Format#secret),
it's recommended that padding is not included in the secret.

Without this change, the tool would error out that the secret is not
base32 if whoever generated the secret followed google's recommendation
there.

With this change, everything else seems to work, so it was just the
validation that bailed, not an underlying library.